Avoiding awful JavaScript behavior regarding new expressions.
Briefly:  new expressions start with the "new" keyword and end with the first
"invocation" expression found, which always ends with arguments listed between
parentheses.  e.g., new tangelo.getPlugin("mongo").Mongo parses as (new
tangelo.getPlugin("mongo")).Mongo, not as new (tangelo.getPlugin("mongo").Mongo)
as you might expect.  The only workarounds to this are to use explicit
parentheses (problematic to rely on users to do, and ugly besides) or to
restructure code not to have such ambiguities between human eyes and the
JavaScript parser.